Product Migrate Create List Edit Update

+++ app/Http/Controllers/AdminPanel/CategoryController.php
@@ -139,7 +139,7 @@ class CategoryController extends Controller
     public function destroy(Category $category, $id)
     {
         $data = Category::find($id);
-        if ($data->image) {
+        if ($data->image && Storage::disk('public')->exists($data->image)) {
             Storage::delete($data->image);
         }
         $data->delete();
